data types has been changed

+++ src/main/java/com/electronic/store/checkout/system/dto/ProductRequest.java
@@ -15,8 +15,8 @@ public class ProductRequest {
 	
 	private String productName;
     private String productDescription;
-    private long price;
-    private long quantity;
+    private Double price;
+    private int quantity;
     private int discount;
